01-naive-rag/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-mongodb>=0.2.0
langchain-community>=0.3.0
pymongo[srv]>=4.6.0
pypdf>=4.0.0
python-dotenv>=1.0.0
certifi>=2024.0.0
rank_bm25>=0.2.2